test: move files in correct dir
